refactor: header state 분리
