+++ web/src/components/Header.tsx
@@ -9,8 +9,8 @@ import { requests } from '../api/request';
 import DarkModeIcon from '../assets/darkmode.webp';
 import LightModeIcon from '../assets/lightmode.webp';
 import { Icons } from '../assets/svg';
-import { setSubPageOpen } from '../reducers/header';
 import { RootState } from '../reducers/index';
+import { setSubPageOpen } from '../types/header';
 import { UserUpdateRequest } from '../types/request';
 import { UserResponse } from '../types/response/User';
 import HeaderModal from './Modal/HeaderModal';
